# server/openapi.yaml
openapi: '3.0.3'
info:
  title: ESA Dono API
  description: |
    Donation platform with Stripe checkout, wallet balance, pledge (smart cart)
    auto-fulfillment, rewards, polls, fund goals, and moderator/admin tooling.

    ## Authentication

    | Method | Mechanism | Used By |
    |--------|-----------|---------|
    | Donor token | `?token=<magic_token>` query param on every request | Donor wallet, spend endpoints, moderator |
    | Admin key | `X-Admin-Key: <ADMIN_API_KEY>` header | Admin routes; also a fallback for moderator routes |
    | Moderator key | `X-Moderator-Key: <MODERATOR_API_KEY>` header | Fallback access to moderator routes, independent of donor roles |
    | Moderator/Admin role | Donor token + `role: MODERATOR` or `role: ADMIN` on donor record | Moderator routes (ADMIN implies MODERATOR access) |
    | Stripe webhook | HMAC signature via `stripe-signature` header | Stripe webhook (skipped when `STRIPE_WEBHOOK_SECRET` unset) |

    ## Rate Limiting

    Spend endpoints (`claim`, `vote`, `custom-entry`, `contribute`) are rate-limited:
    20 requests per minute, keyed by `?token=` or IP. Configurable via `RATE_LIMIT_SPEND`.

    ## Wallet Discount

    When `POST /api/pledge` receives a valid `?token=` (authenticated donor),
    the donor's `balance_remaining` is applied as a discount on the Stripe
    checkout amount. If the wallet covers the entire pledge total, the pledge
...

    ## Money

...
  version: '1.0.0'
servers:
  - url: /api
...
paths:
...
  /pledge:
    post:
      tags: [Pledge]
      summary: Create a pending pledge (smart cart)
      description: |
        Creates a pending pledge from cart items and (when Stripe is configured)
        a Stripe Checkout Session. When a valid donor token is provided via
        `?token=`, the donor's wallet balance is applied as a discount.

        **Streams:** every pledge is routed to exactly one stream via the
        required `stream_id` (used to route the donation to the correct
        overlay). Each cart item's underlying reward/poll/goal must either be
        shared (its own `stream_id` is null) or match the pledge's
        `stream_id` — incentives cannot be mixed across streams in a single
        transaction.

        **Wallet discount flow:**
        - `?token=` present & valid → donor's `balance_remaining` deducted from Stripe charge
        - Wallet ≥ pledge total → pledge fulfilled immediately, redirected to wallet
        - Wallet < pledge total → Stripe charges only `total - discount` (min 50 cents)
        - No `?token=` → no discount applied
      parameters:
        - name: token
          in: query
          required: false
          schema: { type: string }
          description: Optional donor magic token for wallet discount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items, stream_id]
              properties:
                email:
                  type: string
                  format: email
                  description: Donor email (for Stripe pre-fill and magic link delivery)
                comment:
                  type: string
                  maxLength: 500
                stream_id:
                  type: string
                  description: Required. The stream this donation is routed to.
                top_up_cents:
                  type: integer
                  minimum: 0
                  description: Optional additional donation beyond cart items (credited as wallet balance)
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [kind, target_id]
                    properties:
                      kind:
                        type: string
                        enum: [REWARD, POLL_VOTE, GOAL, POLL_CUSTOM]
                      target_id:
                        type: string
                        description: reward_id | poll_option_id | goal_id | poll_id (POLL_CUSTOM)
                      amount_cents:
                        type: integer
                        description: Required for POLL_VOTE, GOAL, POLL_CUSTOM (min 100)
                      poll_id:
                        type: string
                        description: Required for POLL_VOTE and POLL_CUSTOM
                      data:
                        description: Claim data for REWARD, { label } for POLL_CUSTOM
      responses:
        '200':
          description: Pledge created
          content:
            application/json:
              schema:
                type: object
                properties:
                  pledge_token: { type: string }
                  total_cents: { type: integer }
                  expires_at: { type: string, format: date-time }
                  donate_url: { type: string, nullable: true }
                  has_checkout: { type: boolean }
                  wallet_discount_cents: { type: integer, default: 0 }
        '400':
          description: Validation error (invalid items, blocked words, empty cart)
  /pledge/{token}:
    get:
      tags: [Pledge]
      summary: Get pledge status
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Pledge details
          content:
            application/json:
              schema:
                type: object
                properties:
                  pledge_token: { type: string }
                  total_cents: { type: integer }
                  top_up_cents: { type: integer }
                  status:
                    type: string
                    enum: [OPEN, FULFILLED, EXPIRED, CANCELLED]
                  donor_email: { type: string, nullable: true }
                  expires_at: { type: string, format: date-time }
                  magic_token:
                    type: string
                    nullable: true
                    description: Donor's magic token (present when FULFILLED)
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        kind: { type: string }
                        target_id: { type: string }
                        amount_cents: { type: integer }
        '404':
          description: Pledge not found

  # ─── Webhook ───
...
  /admin/pledges:
    get:
      tags: [Admin]
      summary: List all pledges
      security: [{ AdminKey: [] }]
      responses:
        '200':
          description: Pledges with nested items and fulfilled_by donor (capped at 100)
...
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Admin API key matching `ADMIN_API_KEY` env var